Add exec and tests.
This is a free-formed rule that is similar to genrule,
but it is executed when `bazel run`.

This is useful to combine several executables into a
single executable.

The combined script requires the dependant target does
not have `args` and `env`. embedded_exec is used
to wrap args and env.

The rule itself supports 'args' and can be referred to as $@.

Test: TH
Test: bazel test //build/bazel_common_rules/exec/tests
Bug: 228560650
Bug: 229350514
Change-Id: Iadff05a12a98675d9590bc660c07269719da5f53