dist: add support for configurable file permissions
Current code will copy files from Bazel's output
directory to the dist_dir and simultaneously modify
the file permissions to the default umask (644).

If we simply preserve Bazel's output file permissions,
all files have umask 555 due to an upstream Bazel bug:
https://github.com/bazelbuild/bazel/issues/5588.

Some files in the distribution directory should be
executable and others should be readable/writeable.
This adds the ability for the user to choose either
Bazel's default permissions, or the default umask,
as well as the option to apply permissions based on
filename glob-matching.

Bug: 244215515
Change-Id: Ic04cb38a0723221b25c5872cc213818bf08c8941
Signed-off-by: John Moon <quic_johmoo@quicinc.com>