dist: add support for configurable file permissions (second effort)
Current code will copy files from Bazel's output
directory to the dist_dir and simultaneously modify
the file permissions to the default umask (644).

If we simply preserve Bazel's output file permissions,
all files have umask 555 due to an upstream Bazel bug:
https://github.com/bazelbuild/bazel/issues/5588.

Some files in the distribution directory should be
executable and others should be readable/writeable.
This adds the ability for the user to choose
the default umask or the option to apply permissions
based on filename glob-matching.

This reverts commit 0a40cc071e7c64c8483ac529a73562b50db711cd,
aka  relands commit 87e276f7c93bff6d57c882f26161f27b395c1792.
On top of that, this changelist changes it so that
the destination file (which may be write-protected) are
deleted before copying the file.

Bug: 244215515
Test: Run the following 2 times:
      bazel run //common:kernel_dist
Change-Id: I9ff83587324e8928dff20b4c2d5a824ab9053a41
Signed-off-by: John Moon <quic_johmoo@quicinc.com>
Signed-off-by: Yifan Hong <elsk@google.com>